The config param is now a YAML string rather than file name.
For the read_config method, change to taking a YAML string so that a config
string can be injected, for example, in tests.  If the config string is nil, we
still read from the config file.

Now read_config at start of .ydl_files and .load_all methods rather than on
loading of the ydl.rb file.