Allow each Node to refer to the tree's root.
Has been using a global value for the tree root, Ydl.data, which meant that I
could only have a single tree. Also, in resolving cross-references I referred to
this global root, so I could not chain calls on Ydl::Tree. This fixes all that
by keeping a record (the object_id) in each node of the root of the tree that it
is part of and adding method Node#this_tree to dereference the root of the tree.