Fix the problem of instantiating container Nodes.
I was relying on rescuing ArgumentError in instantiate to catch an attempt to
instantiate a node that was really a container node or a parameter node.  This
sets klass only if the Node is meant to be instantiated into that class, and
eliminates the need to catch ArgumentError exceptions in Node#instantiate.

Also, the spec ran 25% faster with this improvement.